+ <le:argument> @optional, @escape + sql\Manager::extract() for one column results + sql:collection type sql\parser\Select : - addJoin() check for recurrences + addClone() for multiple clones + sql\schema table and field interfaces - template\pather::parseArgument() can return tree + sql\template\\Table /name() - template\parser\\Template get namespace from source node + sql\template\\Collection /distinct()

+++ branches/3.0/modules/tester/Parser.php
@@ -79,13 +79,15 @@ class Parser extends tester\Prepare {
       }
     }
 
-    return $this->loadResult($manager, $file, $aArguments);
+    return $this->loadResult($manager, $file, $aArguments, !$test->getx('self:expected', array(), false));
   }
 
-  protected function loadResult($manager, fs\file $file, array $aArguments) {
+  protected function loadResult($manager, fs\file $file, array $aArguments, $bDelete = true) {
+
+    $this->setFile($file);
 
     $result = $manager->load($file, $aArguments, false);
-    $file->delete();
+    if ($bDelete) $file->delete();
 
     return $result;
   }
@@ -122,6 +124,11 @@ class Parser extends tester\Prepare {
     $this->loadResultNode($test);
 
     return parent::test($test, $sContent, $controler, $doc, $file);
+  }
+
+  public function loadScript(array $aArguments = array(), array $aPosts = array(), array $aContexts = array()) {
+
+    return $this->getScriptFile($this->getFile(), $this->buildScriptArguments($aArguments, $aContexts, $aPosts));
   }
 
   public function get($sPath, $bDebug = true) {
